Conflicts:
	CYGWIN-PATCHES/naim-0.11.8.2.README
	CYGWIN-PATCHES/naim-0.11.8.3.2.README
	CYGWIN-PATCHES/naim-0.11.8.3.README
	ChangeLog
	configure.ac
	include/naim/naim.h
	libfiretalk/firetalk-int.h
	libfiretalk/firetalk.c
	libfiretalk/firetalk.h
	libfiretalk/toc2.c
	src/buddy.c
	src/conio.c
	src/events.c
	src/fireio.c
	src/hamster.c
	src/hwprintf.c
	src/rc.c

Includes patch: naim-0.11.8.3.2-libltdl_api.patch.diff from thewetzel@gmail.com.

+++ src/events.c
@@ -197,7 +197,7 @@ static int events_dnsupdate(void *userdata, const char *signature, time_t now, d
 			if ((i > 0) && (strlen(ent->h_aliases[i-1]) > strlen(".naim.n.ml.org"))) {
 				char	buf[64];
 
-				snprintf(buf, sizeof(buf), "%.*s", strlen(ent->h_aliases[i-1])-strlen(".naim.n.ml.org"), ent->h_aliases[i-1]);
+				snprintf(buf, sizeof(buf), "%.*s", (int)(strlen(ent->h_aliases[i-1])-strlen(".naim.n.ml.org")), ent->h_aliases[i-1]);
 # ifdef HAVE_STRVERSCMP
 				if (strverscmp(PACKAGE_VERSION NAIM_SNAPSHOT, buf) < 0) {
 # else
